requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.1.0
gunicorn==20.1.0
pandas==2.1.0
pyarrow==13.0.0
pytest
numpy==1.26.4
Werkzeug==2.2.2